rename NewPassword.js
