+++ src/pages/settings/NewPasswordForm.js
